src/test/kotlin/canon/parser/xml/strategy/SlotMachineStrategyTest.kt
package canon.parser.xml.strategy

import canon.extension.toNode
import canon.parser.xml.CanonXmlParser
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.Assertions.assertNotNull
import org.junit.jupiter.api.Test

class SlotMachineStrategyTest {

    @Test
    fun testParse() {
        val xml = "<slotmachine id='testId' class='testClass' name='testName'>" +
                "</slotmachine>"
        val parsed = SlotMachineStrategy().parse(xml.toNode(), HashMap<String, Any?>(), CanonXmlParser()::toRenderables)

        assertNotNull(parsed)
        assertEquals("testId", parsed.id)
        assertEquals("testClass", parsed.`class`)
        assertEquals("testName", parsed.name)
    }

}
